+++ client/src/features/Event/Overview/components/EventList.tsx
@@ -4,13 +4,14 @@ import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
 import { createInviteCode } from '@/api/mutations/useCreateInviteCode';
+import { Organization } from '@/features/Organization/types/Organization';
 import { Flex } from '@/shared/components/Flex';
 import { Icon } from '@/shared/components/Icon';
 import { Text } from '@/shared/components/Text';
 import { theme } from '@/shared/styles/theme';
 
 import { useModal } from '../../../../shared/hooks/useModal';
-import { Event, Organization } from '../../types/Event';
+import { Event } from '../../types/Event';
 import { EventContainer } from '../containers/EventContainer';
 import { groupEventsByDate } from '../utils/groupEventsByDate';
