+++ client/src/features/Event/Overview/components/OrganizationInfo.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
 import DefaultImage from '@/assets/icon/ahmadda.webp';
-import { Organization } from '@/features/Organization/Select/types/Organization';
+import { Organization } from '@/features/Organization/types/Organization';
 import { Flex } from '@/shared/components/Flex';
 import { Text } from '@/shared/components/Text';
