+++ client/src/features/Organization/Select/components/OrgCard.tsx
@@ -6,7 +6,7 @@ import { Flex } from '@/shared/components/Flex';
 import { Text } from '@/shared/components/Text';
 import { theme } from '@/shared/styles/theme';
 
-import { Organization } from '../types/Organization';
+import { Organization } from '../../types/Organization';
 
 export type OrgCardProps = {
   org: Organization;
